Attempted to refactor and enable mounting of components at the language section. Still missing the package.json etc to actually build the components.
